change face value type of spatial filters to double

+++ processing/src/main/java/org/apache/druid/collections/spatial/search/Bound.java
@@ -42,7 +42,7 @@ public interface Bound
 
   boolean overlaps(ImmutableNode node);
 
-  boolean contains(float[] coords);
+  boolean contains(double[] coords);
 
   Iterable<ImmutablePoint> filter(Iterable<ImmutablePoint> points);
