+++ processing/src/main/java/org/apache/druid/segment/incremental/SpatialDimensionRowTransformer.java
@@ -211,10 +211,10 @@ public class SpatialDimensionRowTransformer implements Function<InputRow, InputR
   }
 
   @Nullable
-  private static Float tryParseFloat(String val)
+  private static Double tryParseFloat(String val)
   {
     try {
-      return Float.parseFloat(val);
+      return Double.parseDouble(val);
     }
     catch (NullPointerException | NumberFormatException e) {
       return null;
@@ -229,17 +229,17 @@ public class SpatialDimensionRowTransformer implements Function<InputRow, InputR
    * @return decoded coordinate, or null if it could not be decoded
    */
   @Nullable
-  public static float[] decode(final String encodedCoordinate)
+  public static double[] decode(final String encodedCoordinate)
   {
     if (encodedCoordinate == null) {
       return null;
     }
 
     final ImmutableList<String> parts = ImmutableList.copyOf(SPLITTER.split(encodedCoordinate));
-    final float[] coordinate = new float[parts.size()];
+    final double[] coordinate = new double[parts.size()];
 
     for (int i = 0; i < coordinate.length; i++) {
-      final Float floatPart = tryParseFloat(parts.get(i));
+      final Double floatPart = tryParseFloat(parts.get(i));
       if (floatPart == null) {
         return null;
       } else {
